+++ src/zenml/zen_server/routers/runs_endpoints.py
@@ -27,6 +27,7 @@ from zenml.constants import (
     VERSION_1,
 )
 from zenml.enums import ExecutionStatus, PermissionType
+from zenml.lineage_graph.lineage_graph import LineageGraph
 from zenml.models import (
     PipelineRunFilterModel,
     PipelineRunResponseModel,
@@ -35,7 +36,6 @@ from zenml.models import (
     StepRunResponseModel,
 )
 from zenml.models.page_model import Page
-from zenml.post_execution.lineage.lineage_graph import LineageGraph
 from zenml.zen_server.auth import AuthContext, authorize
 from zenml.zen_server.exceptions import error_response
 from zenml.zen_server.utils import (
